Move secrets to backend/secrets
