+++ backend/schedulerd/cron_scheduler.go
@@ -6,7 +6,7 @@ import (
 	"github.com/sensu/sensu-go/backend/messaging"
 	"github.com/sensu/sensu-go/backend/store"
 	"github.com/sensu/sensu-go/backend/store/cache"
-	"github.com/sensu/sensu-go/secrets"
+	"github.com/sensu/sensu-go/backend/secrets"
 	"github.com/sirupsen/logrus"
 
 	corev2 "github.com/sensu/sensu-go/api/core/v2"
